Move the account-status wire strings to lib/apierror
- bin-api-manager: Relocate ACCOUNT_EXPIRED / ACCOUNT_DELETED reason,
  message and recovery-endpoint constants plus AccountExpiredDetails from
  lib/middleware to lib/apierror, so the handler layer no longer depends on
  the middleware layer for the vocabulary the two share
- bin-api-manager: Point the v1 gate, POST /auth/login and their tests at the
  new location; the emitted values are unchanged, so the 403 envelopes stay
  byte-identical
