+++ bin-api-manager/lib/middleware/authenticate_test.go
@@ -12,6 +12,7 @@ import (
 	"time"
 
 	amagent "monorepo/bin-agent-manager/models/agent"
+	"monorepo/bin-api-manager/lib/apierror"
 	"monorepo/bin-api-manager/models/auth"
 	modelscommon "monorepo/bin-api-manager/models/common"
 	"monorepo/bin-api-manager/pkg/servicehandler"
@@ -1352,8 +1353,8 @@ func Test_isBlockedAccountStatus(t *testing.T) {
 					if !ok {
 						t.Fatalf("details[0] is not an object: %+v", details[0])
 					}
-					if got, _ := entry["recovery_endpoint"].(string); got != RecoveryEndpointAccountExpired {
-						t.Errorf("Wrong recovery_endpoint. expect: %s, got: %s", RecoveryEndpointAccountExpired, got)
+					if got, _ := entry["recovery_endpoint"].(string); got != apierror.RecoveryEndpointAccountExpired {
+						t.Errorf("Wrong recovery_endpoint. expect: %s, got: %s", apierror.RecoveryEndpointAccountExpired, got)
 					}
 				}
 			}
